backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "beautifulsoup4>=4.13.4",
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "duckduckgo-search>=8.1.1",
    "faiss-cpu>=1.11.0.post1",
    "fastapi[all]>=0.116.1",
    "google-genai>=1.29.0",
    "google-generativeai>=0.8.5",
    "google-search-results>=2.4.2",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.28",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.3",
    "pandas>=2.3.1",
    "playwright>=1.54.0",
    "python-dotenv>=1.1.1",
    "sarvam>=0.0.0",
    "sarvamai>=0.1.18",
    "tavily-python>=0.7.10",
    "twilio>=9.7.0",
    "uvicorn[standard]>=0.35.0",
    "websockets>=15.0.1",
]
